Fix remaining Error Prone findings and gate main with -Werror
Clears the last 12 advisory Error Prone warnings in main sources, then
locks the result in: main now compiles under -Xlint:deprecation -Werror,
so ANY javac or Error Prone WARNING-tier finding fails the build. That
replaces per-check promotion lists, which need maintaining every time
Error Prone ships a new check.

Likely-real defects:
  ThreadLocalUsage - WebService held a non-static ThreadLocal on a
    singleton bean. set/remove were already correctly paired so there was
    no live leak, but static is what the idiom calls for.
  NarrowingCompoundAssignment - totalSplits/totalDividends accumulate
    Stream.count() into an int, silently truncating a long. Widened to
    long rather than casting, so nothing is lost. Callers only log them
    or put them in a Map<String,Object>, and JSON renders long and int
    identically.
  NarrowCalculation - plusMonths(i * 3) in a 0..3 loop cannot overflow,
    but i * 3L costs nothing and says so.

Judgment calls:
  StringSplitter x4 - the check objects to the limit-0 default of
    dropping trailing empty fields, and flags Pattern.split just as much
    as String.split, so hoisting alone does not satisfy it. Each site now
    passes an explicit -1. Checked individually first: all four already
    skip blank entries downstream, so -1 is behaviour-identical and no
    suppression is needed. Patterns are hoisted to static finals too,
    which avoids recompiling them per call in the filing-text hot paths.
  EmptyCatch x3 - the skip-and-continue is deliberate (per-item loops;
    PcapParser has no logger by design, being tuned for throughput). An
    explanatory comment in the catch block satisfies the check, so these
    needed no @SuppressWarnings either.

Test sources stay exempt from -Werror. Their 9 remaining warnings are
structural: 5 UnusedVariable are @Mock fields feeding @InjectMocks, and
3 JavaUtilDate are forced by JJWT's Date-based API. The one free fix
(StringCaseLocaleUsage) is included.

Gate verified by probe, not assumed: a deprecated call and a
NarrowingCompoundAssignment were each injected and each failed the build
with "warnings found and -Werror specified", then reverted.

+++ springboot/src/main/java/com/fattorestreet/sec_api/corporateaction/PriceAdjustmentService.java
@@ -208,8 +208,9 @@ public class PriceAdjustmentService {
                 scheduledDetections.size(), options.force());
 
         int tickersProcessed = 0;
-        int totalSplits = 0;
-        int totalDividends = 0;
+        // long: these accumulate Stream.count(), so an int would hide a lossy narrowing cast.
+        long totalSplits = 0;
+        long totalDividends = 0;
         int totalPricesUpdated = 0;
         int totalSnappedActions = 0;
         int skippedNoAsset = 0;
